Fix position of initial point on y-axis
In the Python version of asciichart, the position of the first tick mark
is sometimes drawn incorrectly due to a rounding error. The value should
be `round` up, but the `int` function is used instead.

Before the change (the tick mark on the y-axis is at 4.5, but it
should be at 6.25):

    > python3 ./pplot --height 3 6 1 8
    >     8.00  ┼
    >     6.25  ┤╮╭
    >     4.50  ┼││
    >     2.75  ┤││
    >     1.00  ┤╰╯

After the change:

    > python3 ./pplot --height 3 6 1 8
    >     8.00  ┼
    >     6.25  ┼╮╭
    >     4.50  ┤││
    >     2.75  ┤││
    >     1.00  ┤╰╯