Add support for NaNs and fix min/max behavior
Data series may now contain NaNs. NaN values will insert a space
in the chart to indicate missing data:

    >>> series = [1,2,3,4,float("nan"),4,3,2,1]
    >>> print(asciichartpy.plot(series))
    4.00  ┤  ╭╴╶╮
    3.00  ┤ ╭╯  ╰╮
    2.00  ┤╭╯    ╰╮
    1.00  ┼╯      ╰

Fix a bug if data values exceeded user-defined minimum and maximum
values. The prior version failed to clamp the values, so the following
would not be possible:

    >>> series = [1,2,3,4,float("nan"),4,3,2,1]
    >>> print(asciichartpy.plot(series, {'minimum': 2, 'maximum': 3}))
    3.00  ┤ ╭─╴╶─╮
    2.00  ┼─╯    ╰─

Improve the API documentation by adding a module docstring as well as
adding several examples to the plot docstring.